+++ modules/components/src/Admin/Dashboard.css
@@ -1,3 +1,38 @@
-div {
-  padding: 4px;
+html,
+body,
+#root,
+.app {
+  height: 100vh;
+  box-sizing: border-box;
+  margin: 0;
+  font-family: Helvetica;
+}
+
+.title-arranger {
+  padding: 10px;
+  background-color: #5b8f94;
+  font-weight: bold;
+  color: white;
+}
+
+.title-elasticsearch {
+  padding: 10px;
+  background-color: #525145;
+  color: white;
+  display: flex;
+  align-items: center;
+  font-size: 11px;
+}
+
+.eshost-input {
+  padding: 0 10px;
+  border: none;
+  outline: none;
+  display: flex;
+  border-bottom: 1px solid #e4e4e4;
+  flex-grow: 1;
+}
+
+.row {
+  display: flex;
 }
