+++ modules/server/package.json
@@ -7,6 +7,7 @@
     "access": "public"
   },
   "scripts": {
+    "start": "nodemon index",
     "prepare": "rm -rf dist&&mkdir dist&&babel src --out-dir dist",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
@@ -26,6 +27,7 @@
   "dependencies": {
     "@arranger/components": "^0.0.1-0.0.11.0",
     "@arranger/mapping-utils": "^0.0.1-0.0.11.0",
+    "@arranger/schema": "^0.0.1-0.0.11.0",
     "@storybook/addon-actions": "^3.3.8",
     "apollo-server-express": "^1.3.1",
     "babel-polyfill": "^6.26.0",
