+++ modules/server/src/graphql.js
@@ -1,8 +1,5 @@
 import express from 'express';
-import bodyParser from 'body-parser';
-import cors from 'cors';
 import { graphqlExpress } from 'apollo-server-express';
-import { rainbow } from 'chalk-animation';
 import elasticsearch from 'elasticsearch';
 
 module.exports = ({
@@ -14,11 +11,8 @@ module.exports = ({
   schema,
   endpoints = ['/graphql', '/graphql/:query'],
 } = {}) => {
-  app.use(cors());
-
   app.use(
     endpoints,
-    bodyParser.json({ limit: '50mb' }),
     schema
       ? graphqlExpress({ schema, context })
       : (req, res) =>
@@ -27,6 +21,4 @@ module.exports = ({
               'schema is undefined. Make sure you provide a valid GraphQL Schema. https://www.apollographql.com/docs/graphql-tools/generate-schema.html',
           }),
   );
-
-  http.listen(port, () => rainbow(`⚡️ Listening on port ${port} ⚡️`));
 };
