+++ package.json
@@ -1,10 +1,11 @@
 {
   "scripts": {
-    "bootstrap": "lerna bootstrap",
+    "bootstrap": "lerna bootstrap --hoist",
     "lerna": "lerna",
     "link": "npm-link-shared modules",
     "publish": "lerna publish",
-    "netlify-build": "npm i && npm run bootstrap -- --scope @arranger/components && cd modules/components && build-storybook"
+    "netlify-build":
+      "npm i && npm run bootstrap -- --scope @arranger/components && cd modules/components && build-storybook"
   },
   "devDependencies": {
     "lerna": "^2.6.0",
