--- sandbox/example-api/.env
@@ -1,32 +1,0 @@
-NODE_PATH=src
-CONFIG=local
-
-API_HOST=http://localhost:5050
-API_PORT=5050
-PORT=3333
-
-GDCAPI=http://localhost:5000
-
-WITH_ES=
-CACHE_MAPPINGS=
-
-ES_HOST=localhost:9200
-ES_TRACE=
-
-ES_GDC_FROM_GRAPH_INDEX=gdc_from_graph
-ES_ANNOTATION_TYPE=annotation
-ES_PROJECT_TYPE=project
-ES_CASE_TYPE=case
-ES_FILE_TYPE=file
-
-ES_CASE_CENTRIC_INDEX=case_centric
-ES_CASE_CENTRIC_TYPE=case_centric
-
-ES_GENE_CENTRIC_INDEX=gene_centric
-ES_GENE_CENTRIC_TYPE=gene_centric
-
-ES_SSM_CENTRIC_INDEX=ssm_centric
-ES_SSM_CENTRIC_TYPE=ssm_centric
-
-ES_SSM_OCC_CENTRIC_INDEX=ssm_occurrence_centric
-ES_SSM_OCC_CENTRIC_TYPE=ssm_occurrence_centric
