--- sandbox/example-api/README.md
@@ -1,6 +1,0 @@
-# Arranger Demo Api
-
-```
-npm i
-WITH_ES=true ES_HOST=http://142.1.177.42:9200 CONFIG=gdc npm run start
-```
