--- sandbox/example-api/hcmi.config.js
@@ -1,13 +1,0 @@
-export let ES_TYPES = {
-  models: {
-    index: 'models',
-    es_type: 'models',
-    name: 'Models',
-    customFields: `
-      cancer_related_somatic_mutations: [String]
-      histopathological_biomarkers: [String]
-    `
-  },
-};
-
-export let ROOT_TYPES = {}
