--- sandbox/example-api/package.json
@@ -1,37 +1,0 @@
-{
-  "name": "example-api",
-  "version": "0.0.1-0.0.11.0",
-  "description": "``` npm i @arranger/server ```",
-  "private": true,
-  "main": "index.js",
-  "scripts": {
-    "start": "nodemon index.js",
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "@arranger/components": "^0.0.1-0.0.11.0",
-    "@arranger/mapping-utils": "^0.0.1-0.0.11.0",
-    "@arranger/schema": "^0.0.1-0.0.11.0",
-    "@arranger/server": "^0.0.1-0.0.11.0",
-    "@storybook/addon-actions": "^3.3.8",
-    "babel-polyfill": "^6.26.0",
-    "dotenv": "^4.0.0",
-    "elasticsearch": "^14.0.0",
-    "express": "^4.16.2",
-    "graphql": "^0.11.7",
-    "node-fetch": "^1.7.3",
-    "socket.io": "^2.0.4",
-    "uuid": "^3.1.0"
-  },
-  "devDependencies": {
-    "@babel/cli": "^7.0.0-beta.35",
-    "@babel/core": "^7.0.0-beta.35",
-    "@babel/plugin-proposal-object-rest-spread": "^7.0.0-beta.35",
-    "@babel/plugin-proposal-optional-chaining": "^7.0.0-beta.35",
-    "@babel/preset-env": "^7.0.0-beta.35",
-    "@babel/register": "^7.0.0-beta.35",
-    "nodemon": "^1.13.3"
-  }
-}
