--- sandbox/example-ui/package.json
@@ -1,27 +1,0 @@
-{
-  "name": "example-ui",
-  "version": "0.0.1-0.0.11.0",
-  "private": true,
-  "dependencies": {
-    "@arranger/components": "^0.0.1-0.0.11.0",
-    "@storybook/addon-actions": "^3.3.8",
-    "babel-polyfill": "^6.26.0",
-    "formik": "^0.10.5",
-    "react": "^16.2.0",
-    "react-dom": "^16.2.0",
-    "react-grid-layout": "^0.16.2",
-    "react-router-dom": "^4.2.2",
-    "react-scripts": "1.0.17",
-    "socket.io-client": "^2.0.4",
-    "uuid": "^3.1.0"
-  },
-  "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test --env=jsdom",
-    "eject": "react-scripts eject"
-  },
-  "devDependencies": {
-    "babel-polyfill": "^6.26.0"
-  }
-}
