--- sandbox/example-ui/src/index.js
@@ -1,7 +1,0 @@
-import './index.css';
-import 'babel-polyfill';
-import React from 'react';
-import ReactDOM from 'react-dom';
-import App from './App';
-
-ReactDOM.render(<App />, document.getElementById('root'));
